requirements.txt
aws-cdk-lib==2.184.1

